Get create and remove directory functions working
These are working, although for both I might want to flesh out the error
handling by checking the value of the errno global variable, which I've
already imported. Also, the remove_directory() function will only remove
an empty directory. I don't think this was mentioned in the assignment
guidelines, so I'll have to ask the prof.