Implement get current directory.
This will retrieve and print out the absolute path to the current
working directory. I'm not sure if this is what he wants, or if I should
truncate it down to just the name of the current node.

Also, I added additional newlines to the menu so the output is nicer and
the different commands have some vertical separation in the terminal.